* Fix build issue (with current nightly) * Build by cargo as library with example
